Single Login Page
 - Removed Multi Login page
 - Using ReturnUrl to redirect User
 - All ref links changed
