+++ Views/Employee/Registration.cshtml
@@ -61,7 +61,7 @@
                         <input type="submit" value="Join" class="btn btn-outline-primary mt-3" />
                     </div>
                 </div>
-                <a class="link-primary text-decoration-none" href="/Employee/Login">Already A Member?</a>
+                <a class="link-primary text-decoration-none" href="/Home/Login">Already A Member?</a>
             </div>
         }
         <div>
