+++ Views/Home/Index.cshtml
@@ -17,7 +17,7 @@
             }
             @if (Session["user"] != null)
             {
-                <a href="/Employee/Logout">Logout</a>
+                <a href="/Home/Logout">Logout</a>
             }
         </div>
         <div class="chart flex-grow-1 p-5">
