+++ Views/Home/Login.cshtml
@@ -10,7 +10,7 @@
             @Html.AntiForgeryToken()
 
             <div class="border p-3 m-auto mt-3">
-                <h4 class="text-center">Serve as Employee</h4>
+                <h4 class="text-center">Login</h4>
                 <p class=" p-3 text-center text-muted">Login to our system and help us fight food waste.</p>
                 @if (TempData["error"] != null)
                 {
