--- Views/Restaurant/Login.cshtml
@@ -1,58 +1,0 @@
-﻿@model WasteFoodDistributionSystem.Models.ViewModel.LoginFormModel
-
-@{
-    ViewBag.Title = "Login";
-}
-<section class="d-flex align-items-center justify-content-between mx-5">
-    <div>
-        @using (Html.BeginForm())
-        {
-            @Html.AntiForgeryToken()
-
-            <div class="border p-3 m-auto mt-3">
-                <h4 class="text-center">Donor Account</h4>
-                <p class=" p-3 text-center text-muted">Don't let good food go to waste, share it and make a difference today!</p>
-                @if (TempData["error"] != null)
-                {
-                    <div class="text-center m-3">
-                        <span class="text-bg-danger p-2">
-                            @TempData["error"]
-                        </span>
-                    </div>
-                }
-                @Html.ValidationSummary(true, "", new { @class = "text-danger" })
-                <div class="form-group">
-                    @Html.LabelFor(model => model.Email, htmlAttributes: new { @class = "control-label col-md-2" })
-                    <div class="col-md-10">
-                        @Html.EditorFor(model => model.Email, new { htmlAttributes = new { @class = "form-control" } })
-                        @Html.ValidationMessageFor(model => model.Email, "", new { @class = "text-danger" })
-                    </div>
-                </div>
-
-                <div class="form-group">
-                    @Html.LabelFor(model => model.Password, htmlAttributes: new { @class = "control-label col-md-2" })
-                    <div class="col-md-10">
-                        @Html.PasswordFor(model => model.Password, new { @class = "form-control" })
-                        @Html.ValidationMessageFor(model => model.Password, "", new { @class = "text-danger" })
-                    </div>
-                </div>
-                <div class="mb-4">
-                    <div class="form-group">
-                        <div class="col-md-offset-2 col-md-10">
-                            <input type="submit" value="Login" class="mt-3 btn btn-outline-primary" />
-                        </div>
-                    </div>
-                </div>
-                <a class="link-info text-decoration-none" href="/Restaurant/Registration">Not a Donor?</a>
-            </div>
-        }
-    </div>
-    <div>
-        <img src="https://thumbs.dreamstime.com/b/reduce-food-waste-cartoon-vector-illustration-good-print-postcard-environmental-warning-sticker-129373112.jpg" alt="Poster" height="600" />
-    </div>
-</section>
-
-
-@section Scripts {
-    @Scripts.Render("~/bundles/jqueryval")
-}
