+++ Views/Restaurant/Registration.cshtml
@@ -61,7 +61,7 @@
                         <input type="submit" value="Join" class="btn btn-outline-primary mt-3" />
                     </div>
                 </div>
-                <a class="link-primary text-decoration-none" href="/Restaurant/Login">Already A Donor?</a>
+                <a class="link-primary text-decoration-none" href="/Home/Login">Already A Donor?</a>
             </div>
         }
         <div>
