+++ Views/Shared/_Layout.cshtml
@@ -23,7 +23,7 @@
             }
             else
             {
-                <a class="btn btn-outline-info px-5" href="/Employee/Login">Help</a>
+                <a class="btn btn-outline-info px-5" href="/Home/Login">Login</a>
             }
         </div>
     </div>
