+++ Views/Shared/_Layout_Dnr.cshtml
@@ -26,7 +26,7 @@
             <a href="@Url.Action("History", "Restaurant")" @(ViewContext.RouteData.Values["Controller"].ToString() == "Restaurant" && ViewContext.RouteData.Values["Action"].ToString() == "History" || ViewContext.RouteData.Values["Action"].ToString() == "DistributerProfile" ? "class=nav-active" : "")><i class="bi bi-clock-history w-25"></i>History</a>
 
             <a href="@Url.Action("DonorProfile", "Restaurant")" @(ViewContext.RouteData.Values["Controller"].ToString() == "Restaurant" && ViewContext.RouteData.Values["Action"].ToString() == "DonorProfile" || ViewContext.RouteData.Values["Action"].ToString() == "Setting" ? "class=nav-active" : "")><i class="bi bi-person-lines-fill w-25"></i>Profile</a>
-            <a href="/Restaurant/Logout"><i class="bi bi-box-arrow-left w-25"></i>Logout</a>
+            <a href="/Home/Logout"><i class="bi bi-box-arrow-left w-25"></i>Logout</a>
         </nav>
         <section class="ds-content">
             @RenderBody()
