+++ WasteFoodDistributionSystem.csproj
@@ -269,6 +269,7 @@
     <Content Include="Views\Employee\FoodDetails.cshtml" />
     <Content Include="Views\Restaurant\DonationDetails.cshtml" />
     <Content Include="Views\Charts\_DonatinoLineChart.cshtml" />
+    <Content Include="Views\Home\Login.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
