+++ Web.config
@@ -16,7 +16,7 @@
   </appSettings>
   <system.web>
 	  <authentication mode="Forms">
-		  <forms name="Auth" loginUrl="/Home/Index"></forms>
+		  <forms name="Auth" loginUrl="/Home/Login"></forms>
 	  </authentication>
     <compilation debug="true" targetFramework="4.8.1" />
     <httpRuntime targetFramework="4.8.1" />
